+++ Examples/Cxx/CreateCMYKImage.cxx
@@ -56,17 +56,16 @@ int main(int argc, char *argv[])
   image.SetPhotometricInterpretation( pi );
   image.SetTransferSyntax( gdcm::TransferSyntax::ExplicitVRLittleEndian );
 
-      gdcm::DataElement pixeldata( gdcm::Tag(0x7fe0,0x0010) );
-      pixeldata.SetByteValue( buf, len );
-      image.SetDataElement( pixeldata );
+  gdcm::DataElement pixeldata( gdcm::Tag(0x7fe0,0x0010) );
+  pixeldata.SetByteValue( buf, (uint32_t)len );
+  image.SetDataElement( pixeldata );
 
-      writer.SetFileName( outfilename );
-      if( !writer.Write() )
-        {
-        return 1;
-        }
-      delete[] buf;
-
+  writer.SetFileName( outfilename );
+  if( !writer.Write() )
+    {
+    return 1;
+    }
+  delete[] buf;
 
   return 0;
 }
